Refactor post-processing
