+++ tools/train_ensemble.py
@@ -19,6 +19,7 @@ from pytorch_lightning import Trainer, seed_everything
 
 from anomalib.config import get_configurable_parameters
 from anomalib.models import get_model
+from anomalib.models.ensemble.post_processing import post_process
 from anomalib.utils.loggers import configure_logger, get_experiment_logger
 
 from anomalib.models.ensemble.ensemble_tiler import EnsembleTiler
@@ -27,10 +28,9 @@ from anomalib.models.ensemble.ensemble_functions import (
     get_ensemble_callbacks,
     prepare_ensemble_configurable_parameters,
     get_prediction_storage,
-    post_process,
 )
 
-from anomalib.models.ensemble.ensemble_metrics import log_metrics
+from anomalib.models.ensemble.post_processing.metrics import log_metrics
 
 logger = logging.getLogger("anomalib")
